requirements.txt
Flask
pillow
pymysql
passlib
gunicorn
